HTML, inline CSS. AMP, Liquid scripts, Handlebars
SMART 8 Email Template HTML and inline CSS for Sales Force Marketing Cloud and Adobe Experience Platform that can be used in the Journeys for marketing purposes! Also for this project, I have used JSON data for personalization in the Jorneys, using AMP and Liquid scripts. The build is also done with the Handlebars script.